Changed the conversion to the UI IStage to fit the UI expectation

+++ src/Angor/Avalonia/Angor.Contexts.Funding/Projects/Application/Dtos/StageDto.cs
@@ -5,5 +5,5 @@ public class StageDto
     public DateTime ReleaseDate { get; set; }
     public long Amount { get; set; }
     public int Index { get; set; }
-    public double RatioOfTotal { get; set; }
+    public decimal RatioOfTotal { get; set; }
 }
